requirements-api.txt
# Heart of Virtue API - Production dependencies only
# For development and testing, use: pip install -r requirements-dev.txt

# Flask and extensions
Flask==3.1.2
Flask-CORS==6.0.2
Flask-SocketIO==5.5.1

# Real-time communication
python-socketio==5.15.1
python-engineio==4.12.3

# Production WSGI server (use with wsgi.py + Procfile)
# eventlet is required for WebSocket support with Flask-SocketIO in production
gunicorn==23.0.0
eventlet==0.40.0

# Additional utilities
python-dotenv==1.2.1
libsql-client==0.3.1
argon2-cffi==23.1.0
cryptography==42.0.5
asgiref==3.11.1
